Minor changes per PR 98.
 - function moved to 'helpers'; use 'new' set literal syntax

https://trello.com/c/OpWI068M/380-after-g9-go-live-removal-of-old-filters-from-search-api-mapping
