+++ tests/app/services/test_query_builder.py
@@ -10,26 +10,10 @@ from app.main.services.query_builder import (
     filter_clause
 )
 from werkzeug.datastructures import MultiDict
+from tests.app.helpers import build_query_params
 
 
 pytestmark = pytest.mark.usefixtures("services_mapping")
-
-
-def build_query_params(keywords=None, page=None, filters=None):
-    query_params = MultiDict()
-    if keywords:
-        query_params["q"] = keywords
-    if filters:
-        for filter_raw_name, filter_value in filters.items():
-            filter_name = "filter_{}".format(filter_raw_name)
-            if isinstance(filter_value, list):
-                for value in filter_value:
-                    query_params.add(filter_name, value)
-            else:
-                query_params[filter_name] = filter_value
-    if page:
-        query_params["page"] = page
-    return query_params
 
 
 def test_should_have_correct_root_element():
